最初と最後に実行されるようファイル名を変更
